# [inductor pattern][cpu] add a new int8 woq mm pattern