--- a/torch/_inductor/fx_passes/quantization.py
+++ b/torch/_inductor/fx_passes/quantization.py
@@ -1449,7 +1449,7 @@ def _is_valid_woq_optimization_pattern():
             # with x.type=bfloat16 and w.type=int8
             x.dtype == torch.bfloat16
             and weight.dtype == torch.int8
-            and scales.dtype == torch.bfloat16
+            and scales.dtype in [torch.bfloat16, torch.float32]
             # _weight_int8pack_mm kernel only supports cpu now
             # TODO: add cuda kernel support instead of calling mul+sum
             and x.device.type == "cpu"
@@ -1460,7 +1460,7 @@ def _is_valid_woq_optimization_pattern():
     return fn
 
 
-def _register_woq_lowering(pattern, computation_woq, computation_reshape):
+def _register_woq_lowering(pattern, computation_woq, computation_reshape=None):
     @register_lowering_pattern(
         pattern,
         extra_check=_is_valid_woq_optimization_pattern(),
@@ -1471,15 +1471,23 @@ def _register_woq_lowering(pattern, computation_woq, computation_reshape):
         scales = kwargs["scales"]
         counters["inductor"]["woq_matcher_count"] += 1
         counters["inductor"]["woq_matcher_nodes"] += len(match.nodes)
-        out_features = weight.get_size()[0]
-        origin_x_size = x.get_size()
-        x_shape = [-1, origin_x_size[-1]]
-        out_shape = origin_x_size[:-1] + [
-            out_features,
-        ]
-        func1 = L[computation_reshape](x, x_shape)
-        func2 = L[computation_woq](func1, weight, scales)
-        return L[computation_reshape](func2, out_shape)
+        if scales.dtype == torch.float32:
+            scales = L[prims.convert_element_type.default](scales, torch.bfloat16)
+        if computation_reshape:
+            out_features = weight.get_size()[0]
+            origin_x_size = x.get_size()
+            x_shape = [-1, origin_x_size[-1]]
+            out_shape = origin_x_size[:-1] + [
+                out_features,
+            ]
+            func1 = L[computation_reshape](x, x_shape)
+            func2 = L[computation_woq](func1, weight, scales)
+            func = L[computation_reshape](func2, out_shape)
+        else:
+            func = L[computation_woq](x, weight, scales)
+        if scales.dtype == torch.float32:
+            func = L[prims.convert_element_type.default](func, torch.float32)
+        return func
 
     return woq
 
@@ -1559,6 +1567,29 @@ def _register_woq_mm_int8_pattern3():
     _register_woq_lowering(_woq_pattern, aten._weight_int8pack_mm.default, aten.reshape)
 
 
+def _register_woq_mm_int8_pattern4():
+    # F.linear(x, weight.to(dtype=x.dtype)) * scales
+    # case of dispatching to mm, w/o output&x reshape
+    _woq_pattern = CallFunction(
+        aten.mul.Tensor,
+        CallFunction(
+            aten.mm.default,
+            KeywordArg("x"),
+            CallFunction(
+                prims.convert_element_type.default,
+                CallFunction(
+                    aten.permute.default,
+                    KeywordArg("weight"),
+                    Arg(),
+                ),
+                Arg(),
+            ),
+        ),
+        KeywordArg("scales"),
+    )
+    _register_woq_lowering(_woq_pattern, aten._weight_int8pack_mm.default)
+
+
 def _register_quantization_lowerings():
     _register_quantization_unary_fusion()
     _register_quantization_binary_fusion()
@@ -1571,6 +1602,7 @@ def _register_woq_lowerings():
     _register_woq_mm_int8_pattern1()
     _register_woq_mm_int8_pattern2()
     _register_woq_mm_int8_pattern3()
+    _register_woq_mm_int8_pattern4()
 
 
 def _is_valid_dequant_promotion_pattern(dtype=torch.float32):