Add int8 woq mm pattern for llama, which successfully hits all the 675 woq linears.

### Implementation
Differences with previous patterns:
- scale is fp32 instead of bf16.
- no reshape for x and output.

### Performance
Performance data on llama, with 1 numa node, freezing mode:
- Before pattern match:
---------- Summary: ----------
inference-latency: 4.810 sec.
first-token-latency: 0.280 sec.
rest-token-latency: 0.146 sec.
P90-rest-token-latency: 0.148 sec.

- After pattern match:
---------- Summary: ----------
inference-latency: 2.537 sec.
first-token-latency: 0.964 sec.
rest-token-latency: 0.051 sec.
P90-rest-token-latency: 0.052 sec.

cc @voznesenskym @penguinwu @EikanWang @jgong5 @Guobing-Chen @XiaobingSuper @zhuhaozhe @blzheng @wenzhe-nrv @jiayisunx @ipiszy @yf225 @chenyang78 @kadeng @muchulee8 @ColinPeppler @amjames @desertfire @chauhang @aakhundov