@leslie-fang-intel 2024-07-22
Why the first token latency even worse after this PR?