@Valentine233 2024-07-25
Summary
1. The fp32-related discussions above are expected to be solved by https://github.com/pytorch/ao/pull/534.2. 
2. The regression of first token needs a further analysis.