@sanchitintel 2024-07-27
I wonder if it'd make sense to have 2 WoQ GEMM auto-tuning templates & micro-kernel in the future that could leverage such patterns - one for QAT (weights not frozen, so won't be cached),  and another that caches weights.