@sanchitintel 2024-08-05
Hi @Valentine233, can you please add info about the precise benchmark you used, perhaps from your bash history?
I used the following command, but [the benchmark](https://github.com/intel/ai-reference-models/blob/main/models_v2/pytorch/llama/inference/cpu/run_llm.py) goes on for a long time, and only posts E2E latency. I'm guessing computing per-token latency requires modifying the transformers library. Please confirm, thanks!

```
 numactl --membind=0 --cpunodebind=0 -C 0-31 python run_llm.py -m "meta-llama/Llama-2-7b-hf" --device cpu --dtype bf16 --output_dir$(pwd)/tmp --batch-size 256 --ws-total-cores 32 --ws-cores-per-instance 32 --weight-dtype INT8 --benchmark --torchao --inductor --num-iter 10 --num-warmup 1 --profile
```

I'm also seeing this warning. Please advise if it's normal. Thanks!
```
W0804 17:20:34.220000 2951501 torch/_dynamo/convert_frame.py:828] [0/8] torch._dynamo hit config.cache_size_limit (8)
W0804 17:20:34.220000 2951501 torch/_dynamo/convert_frame.py:828] [0/8]    function: 'forward' (transformers/src/transformers/models/llama/modeling_llama.py:639)
W0804 17:20:34.220000 2951501 torch/_dynamo/convert_frame.py:828] [0/8]    last reason: 0/0: ___check_obj_id(L['past_key_values'], 94511674602464)
W0804 17:20:34.220000 2951501 torch/_dynamo/convert_frame.py:828] [0/8] To log all recompilation reasons, use TORCH_LOGS="recompiles".
W0804 17:20:34.220000 2951501 torch/_dynamo/convert_frame.py:828] [0/8] To diagnose recompilation issues, see https://pytorch.org/docs/main/torch.compiler_troubleshooting.html.
```