@sanchitintel 2024-08-05
After offline discussion with @Valentine233, it turns out that the runtime error I'm encountering at my end (`torch._dynamo.exc.InternalTorchDynamoError: 'PlainAQTLayout' object has no attribute 'layout_type') is similar to https://github.com/pytorch/ao/pull/534#issuecomment-2246726387.

@Valentine233, is it possible for you to rebase your local PyTorch repo & verify whether or not you're also encountering the same problem? Thanks