@Valentine233 2024-08-08
**Update for first token regression:**

First token has a regression because `aten::_weight_int8pack_mm` is slower than the decomposed one, given certain input shapes. cc @mingfeima @jgong5 @leslie-fang-intel 

### First token
For the first token, the input shapes are x: [128, 4096], w: [4096, 4096], scale: [4096].
**Profiling**
With woq pattern:
![image](https://github.com/user-attachments/assets/18803301-938a-4318-b475-651a89fe083c)
Without woq pattern:
![image](https://github.com/user-attachments/assets/1886de9c-ea07-4ca2-8021-1f693db674d7)


### Next token
For the next token, the input shapes are x: [4, 4096], w: [4096, 4096], scale: [4096].
**Profiling**
With woq pattern:
![image](https://github.com/user-attachments/assets/ff66dd14-a8b5-4dd3-b935-cadbbfdd03fe)
Without woq pattern:
![image](https://github.com/user-attachments/assets/5472b371-1ffb-4096-906d-9ec3f8270c54)

### Reproduce
You could reproduce the result by running the UT added in this PR.