@sanchitintel 2024-08-08
Auto-tuning with WoQ GEMM would not run into this issue because for the first token, the WoQ AMX micro-kernel based implementation would be 1.5x faster than its ATen kernel counterpart (benchmarked with 32 physical cores on a Xeon 4th gen machine with Intel OpenMP & libtamlloc preloaded) .