@sanchitintel 2024-08-08
While auto-tuning would be 1.5x faster for the first token, it too would've resulted in a regression.
`x: [128, 4096], w: [4096, 4096], scale: [4096]` seems to do better if weights are dequantized upfront.