@sanchitintel 2024-08-19
Potential fix - since Xeon CPUs are used on system with a large amount of RAM, we can _optionally_ cache both quantized & dequantized weights for large values of `M` (based on some heuristic), and lower the corresponding FX pattern to a custom function (like the way `quantized_decomposed` custom functions have been defined) that accepts both quantized & dequantized weights. Then that custom function could be lowered to auto-tune WoQ GEMM.

In such a case, the auto-tuning GEMM template could allow a fallback that could be used for large values of `M`, and use the cached dequantized weights if a large value of `M` would be encountered at runtime.